# packages/contracts/identity.openapi.yaml
openapi: 3.1.0
info:
  title: Project Taleem — Identity, Consent & Child-Safe Sign-In API
  version: 0.1.0
  description: >-
    Guardian accounts, verifiable parental consent, learner enrolment, and the child-safe sign-in
    journey (FD-14). Production token cryptography (EdDSA + rotating JWKS) already existed; this
    surface is the journey on top of it.


    Child-safety properties this contract commits to:


    * A learner has no email address, no phone number, and no recoverable credential. A child is
      created by a guardian and signs in with a family code, their own display name, and a short
      PIN.
    * `POST /v1/identity/learners:signin` denies sign-in unless live, un-withdrawn, current-policy
      guardian consent exists. Withdrawal takes effect on the next attempt.
    * Every sign-in failure returns one identical `401` regardless of cause, so no endpoint here is
      an enumeration oracle. `learners:roster` returns an empty list for an unknown family code.
    * Consent and audit are append-only. A withdrawal is a new record; nothing is edited or deleted.
    * Guardian-scoped routes derive the acting guardian from the verified token subject, never from
      the request body. A learner belonging to another family is reported `404`, not `403`, so a
      probe cannot confirm the ref names a real child elsewhere.
    * The `learners:*` routes are halted by the operator kill switch; guardian routes are not, so a
...
  license:
...
servers:
...
paths:
  /v1/identity/guardians:
    post:
      tags: [identity]
      operationId: registerGuardian
      summary: Register a guardian account
      security: []  # public: this is how a caller obtains a token
      description: >-
        Public. Creates an adult account with a family code and returns a first session. Rate
        limited per client.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GuardianRegistration" }
      responses:
        "201":
          description: Account created; session issued
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GuardianSession" }
        "409":
          description: An account already exists for this email address
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "422":
          description: Validation failed (email shape, passphrase policy, display name)
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "429":
          description: Rate limited
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        default:
          description: Error
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
  /v1/identity/guardians:signin:
    post:
      tags: [identity]
      operationId: signInGuardian
      summary: Sign in a guardian
      security: []  # public: this is how a caller obtains a token
      description: >-
        Public. An unknown address and a wrong passphrase return the same `401` with the same
        detail, and take comparable time.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GuardianCredentials" }
      responses:
        "200":
          description: Session issued
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GuardianSession" }
        "401":
          description: Sign-in details are not correct (uniform for every cause)
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "429":
          description: Rate limited
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        default:
          description: Error
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
  /v1/identity/learners:roster:
    post:
      tags: [identity]
      operationId: getFamilyRoster
      summary: Display names for one family code, so a child taps rather than types
      security: []  # public: guarded by the family code, rate limiting, and the kill switch
      description: >-
        Public, rate limited. A `POST` because the family code is a shared secret and must not land
        in a URL, a proxy log, or browser history. Returns display names and grade levels only — no
        refs, no progress. An unknown code returns an empty list, indistinguishable from a family
        with no learners. Halted by the kill switch.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FamilyCodeRequest" }
      responses:
        "200":
          description: Roster (possibly empty)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RosterResponse" }
        "429":
          description: Rate limited
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "503":
          description: Halted by the operator kill switch
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        default:
          description: Error
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
  /v1/identity/learners:signin:
    post:
      tags: [identity]
      operationId: signInLearner
      summary: Sign in a learner (family code + display name + PIN)
      security: []  # public: this is how a learner obtains a token
      description: >-
        Public, rate limited, and gated on live guardian consent. Five wrong PINs lock the learner
        account durably; the owning guardian unlocks it by resetting the PIN. Halted by the kill
        switch.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LearnerCredentials" }
      responses:
        "200":
          description: Session issued
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LearnerSession" }
        "401":
          description: >-
            Sign-in details are not correct. Returned identically for an unknown family code, an
            unknown learner, a wrong PIN, and a locked account.
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "403":
          description: >-
            `CONSENT_REQUIRED` — the credential was correct but no current guardian consent covers
            this learner. Distinguished from `401` on purpose: the caller already holds the family
            code, the name, and the PIN, so this leaks nothing, and it is what tells the child to
            hand the device to their guardian instead of retrying a correct PIN.
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "429":
          description: Rate limited
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "503":
          description: Halted by the operator kill switch
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        default:
          description: Error
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
  /v1/identity/sessions:refresh:
    post:
      tags: [identity]
      operationId: refreshSession
      summary: Exchange a rotating refresh token for a fresh access token
      security: []  # public: the refresh token is itself the credential
      description: >-
        Session continuity that never outlives permission. Refreshing re-runs every gate that
        applied at sign-in — account status and, for a learner, the derived consent state — so a
        withdrawn consent, a locked account, or a reset PIN stops the session here rather than
        thirty days later.


        The token is opaque, single-use, and rotated: the response carries a successor and the
        presented token is consumed. Presenting a consumed token is treated as theft (the legitimate
        holder and a thief hold members of the same chain and are indistinguishable) and revokes the
        entire rotation family. A token bound to a device is refused from any other.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RefreshRequest" }
      responses:
        "200":
          description: New access token and successor refresh token
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RefreshedSession" }
        "401":
          description: >-
            Uniform sign-in failure — unknown token, wrong secret, expired, revoked, wrong device,
            or a consumed token being replayed.
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "403":
          description: "`CONSENT_REQUIRED` — consent lapsed while the session was live"
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "429":
          description: Rate limited
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        default:
          description: Error
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
  /v1/identity/sessions:signout:
    post:
      tags: [identity]
      operationId: signOut
      summary: End a session
      security: []  # public: the refresh token is itself the credential
      description: >-
        Revokes the presented token's whole rotation family. Idempotent, and deliberately identical
        for a token that never existed — signing out must not become a way to probe token validity.
        Signing out on one device does not end sessions on another, because each sign-in starts its
        own family.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SignOutRequest" }
      responses:
        "200":
          description: Signed out (always, whatever was presented)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SignOutResult" }
        "422":
          description: Malformed request body
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        default:
          description: Error
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
  /v1/identity/policy:
    get:
      tags: [identity]
      operationId: getConsentPolicy
      summary: The consent scope vocabulary and current policy version
      security: []  # public: contains no account data
      description: Public — contains no account data. Clients render the consent form from this.
      responses:
        "200":
          description: Policy descriptor
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PolicyDescriptor" }
        default:
          description: Error
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
  /v1/identity/me:
    get:
      tags: [identity]
      operationId: getGuardianIdentityProfile
      summary: The acting guardian, their family code, and their learners
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Guardian profile with per-learner consent state
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GuardianProfile" }
        "401":
          description: Unauthorized
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "403":
          description: Forbidden (not a guardian token)
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "404":
          description: No such guardian account
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        default:
          description: Error
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
  /v1/identity/learners:
    get:
      tags: [identity]
      operationId: listLearners
      summary: The acting guardian's own learners
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Learners with consent state
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LearnerList" }
        "401":
          description: Unauthorized
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "403":
          description: Forbidden
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        default:
          description: Error
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
    post:
      tags: [identity]
      operationId: enrolLearner
      summary: Enrol a child into the acting guardian's family
      description: >-
        A newly enrolled learner cannot sign in until the guardian grants consent — enrol, then
        consent, then learn. Display names are unique within a family because the name is the
        child's sign-in handle there.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LearnerEnrolment" }
      responses:
        "201":
          description: Learner enrolled (cannot sign in until consent is granted)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Learner" }
        "401":
          description: Unauthorized
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "403":
          description: Forbidden
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "422":
          description: PIN policy, duplicate display name, or family size limit
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        default:
          description: Error
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
  /v1/identity/learners/{studentRef}/pin:reset:
    post:
      tags: [identity]
      operationId: resetLearnerPin
      summary: Reset a learner's PIN (also unlocks a locked learner)
      description: >-
        The only recovery path for a child, because a child has no address a reset link could be
        sent to. Owner-only; another family's learner is `404`.
      security: [{ bearerAuth: [] }]
      parameters:
        - name: studentRef
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PinReset" }
      responses:
        "200":
          description: PIN reset; learner reactivated if it was locked
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Learner" }
        "401":
          description: Unauthorized
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "403":
          description: Forbidden
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "404":
          description: Not this guardian's learner
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "422":
          description: PIN policy
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        default:
          description: Error
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
  /v1/identity/consents:
    post:
      tags: [identity]
      operationId: grantConsent
      summary: Grant guardian consent for one learner
      description: >-
        Appends an immutable consent record carrying the policy version and the evidence of capture
        (method, hashed client context, the attestation sentence). Must include `learning_data`.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ConsentGrant" }
      responses:
        "200":
          description: Derived consent state after the grant
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConsentState" }
        "401":
          description: Unauthorized
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "403":
          description: Forbidden (only a guardian may consent — not an operator)
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "404":
          description: Not this guardian's learner
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "422":
          description: Unknown scope, or the required `learning_data` scope was omitted
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        default:
          description: Error
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
  /v1/identity/consents:withdraw:
    post:
      tags: [identity]
      operationId: withdrawConsent
      summary: Withdraw some or all consent scopes for one learner
      description: >-
        Omitting `scopes` withdraws everything — the guardian's stop button. Takes effect on the
        learner's next sign-in attempt, not when a cached claim expires.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ConsentWithdrawal" }
      responses:
        "200":
          description: Derived consent state after the withdrawal
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConsentState" }
        "401":
          description: Unauthorized
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "403":
          description: Forbidden
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "404":
          description: Not this guardian's learner
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        default:
          description: Error
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
  /v1/identity/consents/{studentRef}:
    get:
      tags: [identity]
      operationId: getConsentState
      summary: Current consent state plus the full append-only history for one learner
      security: [{ bearerAuth: [] }]
      parameters:
        - name: studentRef
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Consent state with history, newest first
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConsentStateWithHistory" }
        "401":
          description: Unauthorized
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "403":
          description: Forbidden
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "404":
          description: Not this guardian's learner
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        default:
          description: Error
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
  /v1/identity/audit:
    get:
      tags: [identity]
      operationId: getIdentityAuditTrail
      summary: Append-only identity audit trail for the acting guardian's own family
      description: >-
        Answers "who signed in as my child, and when". Scoped to the guardian and their own
        learners; refs and hashes only, never names or addresses.
      security: [{ bearerAuth: [] }]
      parameters:
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 500, default: 100 }
      responses:
        "200":
          description: Audit events, newest first
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuditTrail" }
        "401":
          description: Unauthorized
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "403":
          description: Forbidden
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        default:
          description: Error
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    GuardianRegistration:
      type: object
      required: [email, passphrase, displayName]
      properties:
        email: { type: string, maxLength: 254 }
        passphrase: { type: string, minLength: 10, maxLength: 256 }
        displayName: { type: string, minLength: 1, maxLength: 60 }
        locale: { type: string, default: "ur" }
    GuardianCredentials:
      type: object
      required: [email, passphrase]
      properties:
        email: { type: string }
        passphrase: { type: string }
...